redirect to signup if no userprofile detected
